refactor abi and notification deps to libs model
