+++ libs/core/__tests__/utils/checkIconSize.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { getFileSizeBytes } from '../../src/utils/getFileSizeBytes';
+import { getFileSizeBytes } from '../../../model/src/utils/getFileSizeBytes';
 
 describe('checkIconSize', () => {
   it("should return zero if url provided doesn't exist", async () => {
