+++ libs/core/src/index.ts
@@ -1,4 +1,3 @@
-export * as Community from './community';
 export * from './ports';
 export * from './requirements-types';
 export * from './shared-types';
