+++ libs/model/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@hicommonwealth/core",
+  "name": "@hicommonwealth/model",
   "private": "true",
   "version": "0.1.0",
-  "description": "Core Abstractions and Utilities",
+  "description": "Core Model - Server Side",
   "files": [
     "build"
   ],
@@ -14,6 +14,7 @@
     "test": "NODE_ENV=test ts-mocha '__tests__/**/*.spec.ts'"
   },
   "dependencies": {
+    "node-object-hash": "^3.0.0",
     "zod": "^3.22.4"
   }
 }
