+++ libs/model/src/index.ts
@@ -1,0 +1,2 @@
+export * as Community from '../src/community';
+export * from './utils/abi';
