+++ libs/model/src/utils/abi.ts
@@ -1,0 +1,13 @@
+import { AbiType } from '@hicommonwealth/core';
+import { hasher } from 'node-object-hash';
+
+export function hashAbi(abi: AbiType): string {
+  const hashInstance = hasher({
+    coerce: true,
+    sort: true,
+    trim: true,
+    alg: 'sha256',
+    enc: 'hex',
+  });
+  return hashInstance.hash(abi);
+}
