+++ libs/model/src/utils/checkIconSize.ts
