+++ libs/model/src/utils/constants.ts
