+++ libs/model/src/utils/getFileSizeBytes.ts
