+++ packages/commonwealth/server/models/contract_abi.ts
@@ -1,7 +1,7 @@
 import { AbiType } from '@hicommonwealth/core';
+import { hashAbi } from '@hicommonwealth/model';
 import type * as Sequelize from 'sequelize'; // must use "* as" to avoid scope errors
 import type { DataTypes } from 'sequelize';
-import { hashAbi } from '../util/abiValidation';
 import type { ModelInstance, ModelStatic } from './types';
 
 export type ContractAbiAttributes = {
