+++ packages/commonwealth/server/routes/communities/create_community_handler.ts
@@ -1,5 +1,5 @@
 import { AppError } from '@hicommonwealth/adapters';
-import { Community } from '@hicommonwealth/core';
+import { Community } from '@hicommonwealth/model';
 import { MixpanelCommunityCreationEvent } from '../../../shared/analytics/types';
 import {
   CreateCommunityOptions,
