+++ packages/commonwealth/tsconfig.json
@@ -24,6 +24,7 @@
   "references": [
     { "path": "../../libs/adapters" },
     { "path": "../../libs/chains" },
-    { "path": "../../libs/core" }
+    { "path": "../../libs/core" },
+    { "path": "../../libs/model" }
   ]
 }
